+++ R/aggregate_game_stats.R
@@ -17,6 +17,10 @@
 #' \describe{
 #' \item{player_id}{ID of the player. Use this to join to other sources.}
 #' \item{player_name}{Name of the player}
+#' \item{player_display_name}{Full name of the player}
+#' \item{position}{Position of the player}
+#' \item{position_group}{Position group of the player}
+#' \item{headshot_url}{URL to a player headshot image}
 #' \item{games}{The number of games where the player recorded passing, rushing or receiving stats.}
 #' \item{recent_team}{Most recent team player appears in `pbp` with.}
 #' \item{season}{Season if `weekly` is `TRUE`}
@@ -97,7 +101,7 @@ calculate_player_stats <- function(pbp, weekly = FALSE) {
 # Prepare data ------------------------------------------------------------
 
   # load plays with multiple laterals
-  mult_lats <- nflreadr::rds_from_url("https://github.com/mrcaseb/nfl-data/raw/master/data/lateral_yards/multiple_lateral_yards.rds") %>%
+  mult_lats <- nflreadr::rds_from_url("https://github.com/nflverse/nflverse-data/releases/download/misc/multiple_lateral_yards.rds") %>%
     dplyr::mutate(
       season = substr(.data$game_id, 1, 4) %>% as.integer(),
       week = substr(.data$game_id, 6, 7) %>% as.integer()
@@ -157,8 +161,22 @@ calculate_player_stats <- function(pbp, weekly = FALSE) {
     dplyr::select(.data$season, .data$season_type, .data$week) %>%
     dplyr::distinct()
 
-  # load gsis_ids of FBs and RBs for RACR
-  racr_ids <- nflreadr::qs_from_url("https://github.com/nflverse/nflfastR-roster/raw/master/data/nflfastR-RB_ids.qs")
+  # we'll join some player information like position or full name later
+  # so we load it here to be able to use it for racr ids as well
+  player_info <- nflreadr::load_players() %>%
+    dplyr::select(
+      "player_id" = "gsis_id",
+      "player_display_name" = "display_name",
+      "player_name" = "short_name",
+      "position",
+      "position_group",
+      "headshot_url" = "headshot"
+    )
+
+  # load gsis_ids of RBs, FBs and HBs for RACR
+  racr_ids <- player_info %>%
+    dplyr::filter(.data$position %in% c("RB", "FB", "HB")) %>%
+    dplyr::select("gsis_id" = "player_id")
 
 # Passing stats -----------------------------------------------------------
 
@@ -610,6 +628,21 @@ calculate_player_stats <- function(pbp, weekly = FALSE) {
         dplyr::everything()
       )
   }
+
+  # data is missing position and player name can be messed up in pbp
+  # so we join player information next
+  player_df <- player_df %>%
+    dplyr::select(-"player_name") %>%
+    dplyr::left_join(player_info, by = "player_id") %>%
+    dplyr::select(
+      "player_id",
+      "player_name",
+      "player_display_name",
+      "position",
+      "position_group",
+      "headshot_url",
+      dplyr::everything()
+    )
 
   return(player_df)
 }